requirements.txt
python-dotenv>=1.0.0
openai>=1.12.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
